+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
   - Unit selection
+  - Day panel first implementation
+
 
 ## [0.0.4] - 2018-02-11
 ### Added